+++ src/tad/DI52/Bindings/ResolverInterface.php
@@ -57,4 +57,20 @@ interface tad_DI52_Bindings_ResolverInterface
      * Boots up the application calling the `boot` method of each registered service provider.
      */
     public function boot();
+
+    /**
+     * Checks whether if an interface or class has been bound to a concrete implementation.
+     *
+     * @param string $classOrInterface
+     * @return bool
+     */
+    public function isBound($classOrInterface);
+
+    /**
+     * Checks whether a tag group exists in the container.
+     *
+     * @param string $tag
+     * @return bool
+     */
+    public function hasTag($tag);
 }