+++ src/tad/DI52/ServiceProvider.php
@@ -3,9 +3,17 @@
 abstract class tad_DI52_ServiceProvider implements tad_DI52_ServiceProviderInterface
 {
     /**
+     * Whether the service provider will be a deferred one or not.
+     *
+     * @var bool
+     */
+    protected $deferred = false;
+
+    /**
      * @var tad_DI52_Container
      */
     protected $container;
+
 
     /**
      * tad_DI52_ServiceProvider constructor.
@@ -25,4 +33,24 @@ abstract class tad_DI52_ServiceProvider implements tad_DI52_ServiceProviderInter
      * Binds and sets up implementations at boot time.
      */
     abstract public function boot();
+
+    /**
+     * Whether the service provider will be a deferred one or not.
+     *
+     * @return bool
+     */
+    public function isDeferred()
+    {
+        return $this->deferred;
+    }
+
+    /**
+     * Returns an array of the class or interfaces bound and provided by the service provider.
+     *
+     * @return array
+     */
+    public function provides()
+    {
+        return array();
+    }
 }