+++ src/tad/DI52/ServiceProviderInterface.php
@@ -11,4 +11,13 @@ interface tad_DI52_ServiceProviderInterface
      * Binds and sets up implementations at boot time.
      */
     public function boot();
+
+    public function provides();
+
+    /**
+     * Whether the service provider will be a deferred one or not.
+     *
+     * @return bool
+     */
+    public function isDeferred();
 }